more tests parallelized for default config
