+++ test/e2e/data_setup/config/recordset/add.config.json
@@ -1,0 +1,20 @@
+{
+  "catalog": {
+    "acls": [
+      {"name": "content_write_user", "user": "*" },
+      {"name": "content_read_user", "user": "*" }
+    ]
+  },
+  "schema": {
+    "name": "product",
+    "createNew": true,
+    "path": "test/e2e/data_setup/schema/product.json"
+  },
+  "tables": {
+    "createNew": true
+  },
+  "entities": {
+    "createNew": true,
+    "path": "test/e2e/data_setup/data/product"
+  }
+}
