+++ test/e2e/data_setup/config/recordset/add.dev.json
@@ -1,25 +1,8 @@
 {
   "setup": {
-    "schemaConfigurations": [{
-      "catalog": {
-        "acls": [
-          {"name": "content_write_user", "user": "*" },
-          {"name": "content_read_user", "user": "*" }
-        ]
-      },
-      "schema": {
-        "name": "product",
-        "createNew": true,
-        "path": "test/e2e/data_setup/schema/product.json"
-      },
-      "tables": {
-        "createNew": true
-      },
-      "entities": {
-        "createNew": true,
-        "path": "test/e2e/data_setup/data/product"
-      }
-    }],
+    "schemaConfigurations": [
+        "recordset/add.config.json"
+    ],
     "schema": "product"
   },
   "cleanup" : true,
