+++ test/e2e/data_setup/config/search/dev.json
@@ -1,0 +1,8 @@
+{
+    "setup": {
+        "schemaConfigurations": [
+            "search/config.json"
+        ]
+    },
+    "cleanup" : true
+}
