+++ test/e2e/data_setup/config/viewer/config.json
@@ -1,0 +1,20 @@
+{
+  "catalog": {
+    "acls": [
+      {"name": "content_write_user", "user": "*" },
+      {"name": "content_read_user", "user": "*" }
+    ]
+  },
+  "schema": {
+    "name": "product-viewer",
+    "createNew": true,
+    "path": "test/e2e/data_setup/schema/product-viewer.json"
+  },
+  "tables": {
+    "createNew": true
+  },
+  "entities": {
+    "createNew": true,
+    "path": "test/e2e/data_setup/data/product"
+  }
+}
