+++ test/e2e/data_setup/config/viewer/dev.json
@@ -1,0 +1,9 @@
+{
+  "setup": {
+    "schemaConfigurations": [
+        "viewer/config.json"
+    ],
+    "schema": "product-viewer"
+  },
+  "cleanup" : true
+}
